Move Array grouping extension to separate file
Relocated the Array extension for grouping by date key path from Syncable.swift to a new file Array+Grouping.swift for better modularity and code organization.
